+++ NEWS.md
@@ -12,6 +12,7 @@
 
 ### Bug fixes
 
+* Various accessibility improvements for screen readers.
 * `colFormat(percent = TRUE)` now works correctly when viewing tables in IE11.
 * Table searching now works when row selection is enabled.
 * In `colDef()`, R functions for `class` and `style` now always receive cell values as single elements.